Temporary fix for scipy.sparse.linalg.inv
Due to change introduced at scipy 1.11, choice of dtype became sensitive.
For sparse matrix indices, row and col, have to be int64. However, here for
scipy.sparse.linalg.inv (or splu), those can not be int64 (int_) but have to be
int32 (intc). This is not consistent, so this can be changed or fixed
at some versions of scipy.